requirements.txt
Django==5.1.7
mysqlclient==2.2.7
django-crontab==0.7.1
gunicorn==23.0.0
pymemcache==4.0.0
python-dotenv==1.0.0
whitenoise==6.7.0
Pillow==11.1.0
requests==2.32.3
cent==4.0.0
PyJWT==2.10.0
django-cors-headers==4.6.0
Faker==37.1.0
tqdm==4.67.1
asgiref==3.8.1
autoslug==1.0.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.3
idna==3.10
inflection==0.5.1
packaging==25.0
pycparser==2.22
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
unicode-slugify==0.1.5
Unidecode==1.3.8
urllib3==2.4.0